+++ src/models.py
@@ -1,21 +1,29 @@
-import xgboost as xgb
+import numpy as np
+import pandas as pd
+import lightgbm as lgb
 from attrdict import AttrDict
-from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import StandardScaler
+from sklearn.externals import joblib
+from deepsense import neptune
 from steppy.base import BaseTransformer
-from steppy.utils import get_logger
-from toolkit.sklearn_transformers.models import SklearnClassifier
+
+from .utils import get_logger
 
 logger = get_logger()
+ctx = neptune.Context()
 
 
-class XGBoost(BaseTransformer):
-    def __init__(self, **params):
+class LightGBM(BaseTransformer):
+    def __init__(self, model_config, callback_config):
         super().__init__()
-        logger.info('initializing XGBoost...')
-        self.params = params
-        self.training_params = ['nrounds', 'early_stopping_rounds']
+        logger.info('initializing LightGBM...')
+        self.params = model_config
+        self.training_params = ['number_boosting_rounds', 'early_stopping_rounds']
         self.evaluation_function = None
+        if callback_config['run_with_callback']:
+            callback_config.pop('run_with_callback', None)
+            self.callbacks = callbacks(callback_config)
+        else:
+            self.callbacks = None
 
     @property
     def model_config(self):
@@ -28,57 +36,99 @@ class XGBoost(BaseTransformer):
                          if param in self.training_params})
 
     def fit(self,
-            X, y,
-            X_valid, y_valid,
-            feature_names=None,
-            feature_types=None,
+            X,
+            y,
+            X_valid,
+            y_valid,
+            feature_names='auto',
+            categorical_features='auto',
             **kwargs):
-        train = xgb.DMatrix(X,
-                            label=y,
-                            feature_names=feature_names,
-                            feature_types=feature_types)
-        valid = xgb.DMatrix(X_valid,
-                            label=y_valid,
-                            feature_names=feature_names,
-                            feature_types=feature_types)
-
         evaluation_results = {}
-        self.estimator = xgb.train(params=self.model_config,
-                                   dtrain=train,
-                                   evals=[(train, 'train'), (valid, 'valid')],
+
+        self._check_target_shape_and_type(y, 'y')
+        self._check_target_shape_and_type(y_valid, 'y_valid')
+        y = self._format_target(y)
+        y_valid = self._format_target(y_valid)
+
+        logger.info('LightGBM, train data shape        {}'.format(X.shape))
+        logger.info('LightGBM, validation data shape   {}'.format(X_valid.shape))
+        logger.info('LightGBM, train labels shape      {}'.format(y.shape))
+        logger.info('LightGBM, validation labels shape {}'.format(y_valid.shape))
+
+        data_train = lgb.Dataset(data=X,
+                                 label=y,
+                                 feature_name=feature_names,
+                                 categorical_feature=categorical_features,
+                                 **kwargs)
+        data_valid = lgb.Dataset(X_valid,
+                                 label=y_valid,
+                                 feature_name=feature_names,
+                                 categorical_feature=categorical_features,
+                                 **kwargs)
+
+        self.estimator = lgb.train(self.model_config,
+                                   data_train,
+                                   feature_name=feature_names,
+                                   categorical_feature=categorical_features,
+                                   valid_sets=[data_train, data_valid],
+                                   valid_names=['data_train', 'data_valid'],
                                    evals_result=evaluation_results,
-                                   num_boost_round=self.training_config.nrounds,
+                                   num_boost_round=self.training_config.number_boosting_rounds,
                                    early_stopping_rounds=self.training_config.early_stopping_rounds,
                                    verbose_eval=self.model_config.verbose,
-                                   feval=self.evaluation_function)
+                                   feval=self.evaluation_function,
+                                   callbacks=self.callbacks,
+                                   **kwargs)
         return self
 
-    def transform(self, X, y=None, feature_names=None, feature_types=None, **kwargs):
-        X_DMatrix = xgb.DMatrix(X,
-                                label=y,
-                                feature_names=feature_names,
-                                feature_types=feature_types)
-        prediction = self.estimator.predict(X_DMatrix)
+    def transform(self, X, **kwargs):
+        prediction = self.estimator.predict(X)
         return {'prediction': prediction}
 
     def load(self, filepath):
-        self.estimator = xgb.Booster(params=self.model_config)
-        self.estimator.load_model(filepath)
+        self.estimator = joblib.load(filepath)
         return self
 
     def persist(self, filepath):
-        self.estimator.save_model(filepath)
+        joblib.dump(self.estimator, filepath)
+
+    def _check_target_shape_and_type(self, target, name):
+        if not any([isinstance(target, obj_type) for obj_type in [pd.Series, np.ndarray, list]]):
+            raise TypeError(
+                '"{}" must be "numpy.ndarray" or "Pandas.Series" or "list", got {} instead.'.format(type(target)))
+        try:
+            assert len(target.shape) == 1, '"{}" must be 1-D. It is {}-D instead.'.format(name,
+                                                                                          len(target.shape))
+        except AttributeError:
+            print('Cannot determine shape of the {}. '
+                  'Type must be "numpy.ndarray" or "Pandas.Series" or "list", got {} instead'.format(name,
+                                                                                                     type(target)))
+
+    def _format_target(self, target):
+
+        if isinstance(target, pd.Series):
+            return target.values
+        elif isinstance(target, np.ndarray):
+            return target
+        elif isinstance(target, list):
+            return np.array(target)
+        else:
+            raise TypeError(
+                '"{}" must be "numpy.ndarray" or "Pandas.Series" or "list", got {} instead.'.format(type(target)))
 
 
-def get_sklearn_classifier(ClassifierClass, normalize=False, **kwargs):
+def callbacks(callback_config):
+    neptune_monitor = neptune_monitor_lgbm(**callback_config['neptune_monitor'])
+    return [neptune_monitor]
 
-    class SklearnBinaryClassifier(SklearnClassifier):
-        def transform(self, X, y=None, target=1, **kwargs):
-            prediction = self.estimator.predict_proba(X)[:, target]
-            return {SklearnClassifier.RESULT_KEY: prediction}
 
-    if normalize:
-        return SklearnBinaryClassifier(Pipeline([('standarizer', StandardScaler()),
-                                                 ('classifier', ClassifierClass(**kwargs))]))
+def neptune_monitor_lgbm(channel_prefix=''):
+    def callback(env):
+        for name, loss_name, loss_value, _ in env.evaluation_result_list:
+            if channel_prefix != '':
+                channel_name = '{}_{}_{}'.format(channel_prefix, name, loss_name)
+            else:
+                channel_name = '{}_{}'.format(name, loss_name)
+            ctx.channel_send(channel_name, x=env.iteration, y=loss_value)
 
-    return SklearnBinaryClassifier(ClassifierClass(**kwargs))
+    return callback
